WiP allow using repo name from config in URL
some git hosting systems do not have the browse url as simple as
clone URL + some symbols.

For example, with Gerrit and the Gitiles plugin, the clone URL is
https://example.com/my-org/my-project
while the browse URL will be
https://example.com/plugins/gitiles/my-org/my-project/+/refs/head/master

This commit allows to use the 'name' of the repo in the config.json
as substitution variable in the browse_url, which can greatly simplify
configuration for such cases.